edit and delete service integrated

+++ src/app/components/leave_screen/deleteAlert.tsx
@@ -7,6 +7,7 @@ import { AppIcon, snackBarMessage, snackErrorBottom } from '../../common';
 import { dataType } from '../../interface';
 import { RequestContext } from '../../reducer';
 import { deleteRequest, cancelLeave } from '../../services';
+import { deleteTimeLog } from '../../services/timeLogService';
 import { getUser } from '../../utils';
 
 const DeleteAlert = ({
@@ -21,32 +22,39 @@ const DeleteAlert = ({
   const [showAlert, setShowAlert] = useState(false);
   const show = () => setShowAlert(true);
   const hide = () => setShowAlert(false);
-  const { dispatchRequest } = useContext(RequestContext);
 
   const onDelete = async () => {
     const user = await getUser();
-
-    if (other) {
-      if (new Date(item.leave_date.startDate) > new Date()) {
-        cancelLeave(item.id)
-          .then((data) => {
-            dispatchRequest({ type: 'UPDATEQUOTA', payload: data.quota });
-            dispatchRequest({ type: 'CANCEL', payload: data.leave });
-            snackBarMessage('Request Cancelled');
+    if (!timelog) {
+      const { dispatchRequest } = useContext(RequestContext);
+      if (other) {
+        if (new Date(item.leave_date.startDate) > new Date()) {
+          cancelLeave(item.id)
+            .then((data) => {
+              dispatchRequest({ type: 'UPDATEQUOTA', payload: data.quota });
+              dispatchRequest({ type: 'CANCEL', payload: data.leave });
+              snackBarMessage('Request Cancelled');
+            })
+            .catch((err) => console.log(err));
+        } else {
+          snackErrorBottom({ message: 'You cannot cancel request now' });
+        }
+      } else {
+        deleteRequest(item.id)
+          .then(async (data) => {
+            dispatchRequest({ type: 'UPDATEQUOTA', payload: data });
+            dispatchRequest({ type: 'DELETE', payload: item.id });
+            snackBarMessage('Request deleted');
           })
           .catch((err) => console.log(err));
-      } else {
-        snackErrorBottom({ message: 'You cannot cancel request now' });
       }
-    } else {
-      deleteRequest(item.id)
-        .then(async (data) => {
-          dispatchRequest({ type: 'UPDATEQUOTA', payload: data });
-          dispatchRequest({ type: 'DELETE', payload: item.id });
-          snackBarMessage('Request deleted');
-        })
-        .catch((err) => console.log(err));
     }
+  };
+
+  const onLogDelete = () => {
+    deleteTimeLog(item.id)
+      .then((data) => console.log('deleted', data))
+      .catch((err) => console.log(err));
   };
 
   return (
@@ -66,7 +74,9 @@ const DeleteAlert = ({
           <AppIcon name="alert" color={colors.tomato} size={30} />
           <View style={style.main}>
             <Dialog.Title style={style.text1}>
-              {other ? 'Cancel' : 'Delete'} the request ?
+              {timelog
+                ? 'Delete the timelog?'
+                : (other ? 'Cancel' : 'Delete') + ' the request ?'}
             </Dialog.Title>
             <Dialog.Title style={style.text2}>This cant be undone</Dialog.Title>
           </View>
@@ -80,7 +90,7 @@ const DeleteAlert = ({
           <Dialog.Button
             label={other ? 'YES' : 'DELETE'}
             onPress={() => {
-              onDelete();
+              timelog ? onLogDelete() : onDelete();
               hide();
             }}
             style={style.delete}
